typescript(vx-demo): rewrite ZoomI example in TypeScript
